refactor: put most of the JS code in helpers
There is actually one small difference in this version: when previously a validation error would throw synchronously even when a promise was expected, now it returns a rejected promise. The `throw` behavior is still the same when a callback is provided, which is probably still not great as it means you need two handlers: one for anything thrown by the call to `decode` and another for the `err` in the callback.